31/05/19 - Check in #2 Code 90% finished

+++ src/java/org/ExeRiver/model/Observer.java
@@ -16,6 +16,6 @@ public interface Observer {
     //method to update the observer, used by subject
 	public void updateBookDetails(int profR1Rating, int profR2Rating, int editorRating, String profR1Comments, String profR2Comments, boolean overallAcceptStatus, boolean manuscriptAgreed);
         public void updateBookRevisionDetails(String editorComments, String manuscript);
-        public void updatePaymentDetails(float amount, boolean complete);
+        public void updatePaymentDetails(double amount, boolean complete);
         public void updateMeetingDetails(String meetingNotes);
 }
